Moving static/ files to /src/static and refactoring its references
