+++ promova-cultura-vuejs/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <link rel="shortcut icon" href="./static/favicon.ico" />
+    <link rel="shortcut icon" href="src/static/favicon.ico" />
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <title>Promova Cultura</title>
   </head>
