+++ promova-cultura-vuejs/src/App.vue
@@ -12,10 +12,10 @@
 <script>
 import NavbarMenu from "@/components/NavbarMenu";
 
-import "../static/bootstrap/bootstrap.min.js";
-import "../static/jsmaps/jsmaps.js";
-import "../static/jsmaps/jsmaps-panzoom.js";
-import "../static/jsmaps/jsmaps-libs.js";
+import "./static/bootstrap/bootstrap.min.js";
+import "./static/jsmaps/jsmaps.js";
+import "./static/jsmaps/jsmaps-panzoom.js";
+import "./static/jsmaps/jsmaps-libs.js";
 
 export default {
   name: "App",
