+++ promova-cultura-vuejs/src/components/Map/info/ProponentBarCard.vue
@@ -6,18 +6,18 @@
       </div>
       <div class="card-body">
         <div class="card-number">
-          <img src="../../../../static/svg-icons/proponente_verde.svg"
+          <img src="../../../static/svg-icons/proponente_verde.svg"
                class="card-icon"></img>
           {{this.totalProponents}}
         </div>
         <div class="card-number">
-          <img src="../../../../static/svg-icons/check_verde.svg"
+          <img src="../../../static/svg-icons/check_verde.svg"
                class="card-icon"></img>
           R$ {{(this.totalapprovedAmount).toLocaleString('pt-BR', {
           minimumFractionDigits: 2})}}
         </div>
         <div class="card-number">
-          <img src="../../../../static/svg-icons/cifrao_verde.svg"
+          <img src="../../../static/svg-icons/cifrao_verde.svg"
                class="card-icon"></img>
           R$ {{(this.totalraisedAmount).toLocaleString('pt-BR', {
           minimumFractionDigits: 2})}}
