+++ promova-cultura-vuejs/src/static/funnel/chart.js
