+++ promova-cultura-vuejs/src/static/jsmaps/brazil.js
