+++ promova-cultura-vuejs/src/static/jsmaps/jsmaps-panzoom.js
